+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemFireworkStarStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemFireworkStarBaseComponent;
 
-public interface ItemFireworkStarStack extends ItemStack {
+public interface ItemFireworkStarStack extends ItemStack, ItemFireworkStarBaseComponent {
 }