+++ server/src/main/java/org/allaymc/server/entity/type/EntityTypeInitializer.java
@@ -213,4 +213,13 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityLivingComponentImpl::new, EntityLivingComponentImpl.class)
                 .build();
     }
+
+    public static void initFireworkRocket() {
+        EntityTypes.FIREWORKS_ROCKET = AllayEntityType
+                .builder(EntityFireworksRocketImpl.class)
+                .vanillaEntity(EntityId.FIREWORKS_ROCKET)
+                .addComponent(EntityFireworksRocketPhysicsComponentImpl::new, EntityFireworksRocketPhysicsComponentImpl.class)
+                .addComponent(EntityFireworksRocketBaseComponentImpl::new, EntityFireworksRocketBaseComponentImpl.class)
+                .build();
+    }
 }