+++ server/src/main/java/org/allaymc/server/item/impl/ItemFireworkStarStackImpl.java
@@ -1,7 +1,9 @@
 package org.allaymc.server.item.impl;
 
+import lombok.experimental.Delegate;
 import org.allaymc.api.component.Component;
 import org.allaymc.api.item.ItemStackInitInfo;
+import org.allaymc.api.item.component.ItemFireworkStarBaseComponent;
 import org.allaymc.api.item.interfaces.ItemFireworkStarStack;
 import org.allaymc.server.component.ComponentProvider;
 
@@ -11,5 +13,11 @@ public class ItemFireworkStarStackImpl extends ItemStackImpl implements ItemFire
     public ItemFireworkStarStackImpl(ItemStackInitInfo initInfo,
                                      List<ComponentProvider<? extends Component>> componentProviders) {
         super(initInfo, componentProviders);
+    }
+
+    @Delegate
+    @Override
+    protected ItemFireworkStarBaseComponent getBaseComponent() {
+        return (ItemFireworkStarBaseComponent) super.getBaseComponent();
     }
 }