+++ server/src/main/java/org/allaymc/server/item/type/ItemTypeInitializer.java
@@ -882,4 +882,20 @@ public final class ItemTypeInitializer {
                 .vanillaItem(ItemId.PAINTING)
                 .build();
     }
+
+    public static void initFireworkStar() {
+        ItemTypes.FIREWORK_STAR = AllayItemType
+                .builder(ItemFireworkStarStackImpl.class)
+                .vanillaItem(ItemId.FIREWORK_STAR)
+                .addComponent(ItemFireworkStarBaseComponentImpl::new, ItemFireworkStarBaseComponentImpl.class)
+                .build();
+    }
+
+    public static void initFireworkRocket() {
+        ItemTypes.FIREWORK_ROCKET = AllayItemType
+                .builder(ItemFireworkRocketStackImpl.class)
+                .vanillaItem(ItemId.FIREWORK_ROCKET)
+                .addComponent(ItemFireworkRocketBaseComponentImpl::new, ItemFireworkRocketBaseComponentImpl.class)
+                .build();
+    }
 }